Fix trace() tests for uint64_t
The trace() tests for uint64_t were using (uint64_t)-1 which is not
quite resulting in the correct value.  It does not yield a 64-bit
representation of -1 but rather takes the 32-bit representation of
-1 and converts it into a 64-bit value with 0s in the 32-bit higher
bits.  The tests used to pass because of the bug that was fixed
recently (c61ae825 'Fix DT_TOK_INT for values between INT32_MAX and
UINT32_MAX').

By using -1ul we ensure that a true 64-bit representation of -1 is
used in these tests.

+++ test/unittest/actions/trace/tst.uint64_t-quiet.d
@@ -16,6 +16,6 @@
 
 BEGIN
 {
-	trace((uint64_t)-1);
+	trace(-1ul);
 	exit(0);
 }
